*Non-Boussinesq revision of MOM_CVMix_KPP
  This commit revises MOM_CVMix_KPP to work in an appropriate mixture of
thickness and vertical extent variables to enable their use in non-Boussinesq
mode, using thickness_to_dz to convert between the two, while retaining the
previous answers in Boussinesq mode.  This includes the use of a layer
thicknesses rather than a vertical distance in the denominator of the
calculation of the buoyancy frequency and the replacement of the layer thickness
argument (h) to KPP_smooth_BLD with a layer vertical extent (dz).  When in
non-Boussinesq mode, the buoyancy difference between layers is normalized by the
average of the density of the two layers rather than the Boussinesq reference
density. This commit eliminates any direct or indirect dependency on the
Boussinesq reference density in CVMix_KPP when in non-Boussinesq mode.

  This set of changes includes changing the units of 1 internal variables and
the addition of 2 new internal variables, and a change to the name and units of
one argument to KPP_smooth_BLD.  These changes lead to the removal of 5 factors
of GV%H_to_Z.

  Answers will change in non-Boussinesq mode when USE_KPP is true, but they are
bitwise identical in all Boussinesq test cases.