+*Non-Boussinesq revision of lateral_mixing_coeffs
  This commit revises lateral_mixing_coeffs to work in an appropriate mixture of
thickness and vertical extent variables to avoid any dependence on the
Boussinesq reference density in non-Boussinesq mode, while retaining the
previous answers in Boussinesq mode.

  This commit adds the new runtime parameter FULL_DEPTH_EADY_GROWTH_RATE to
indicate that the denominator of an Eady growth rate calculation should be based
on the full depth of the water column, rather than the nominal depth of the
bathymetry.  The new option is only the default for fully non-Boussinesq cases.

  A primordial horizontal indexing bug was corrected in the v-direction slope
calculation.  Because it only applies for very shallow bathymetry, does not
appear to impact any existing test cases and went undetected for at least 12
years, it was corrected directly rather than wrapping in another new runtime
flag.  However, this bug is being retained for now in a comment to help with
review and debugging if the answers should change unexpectedly in some yet-to-be
identified configuration.

  Two debugging checksums were added for the output variables calculated in
calc_resoln_function.

  The case of some indices was corrected to follow the MOM6 soft convention using
case to indicate the staggering position of variables.  The previously incorrect
units of one comment were also fixed.  There is a new logical element in the
VarMix_CS type.  To accommodate these changes there are three new internal
variables in calc_slope_functions_using_just_e.  A total of 9 GV%H_to_Z
conversion factors were eliminated with this commit.

  N2 is no longer calculated separately in calc_slope_functions_using_just_e,
but this code is left in a comment as it may be instructive. This commit
involved changing the units of one internal variable in calc_QG_Leith_viscosity
to use inverse thickness units (as its descriptive comment already indicated).
There are already known problems with calc_QG_Leith_viscosity as documented with
a fatal error; this will be addressed in a subsequent commit.

  All answers are bitwise identical in the existing MOM6-examples test suite,
but they will change when fully non-Boussinesq, and there is a new entry in some
MOM_parameter_doc files.