+Remove build_grid_arbitrary
  Removed the unused (and unusable) routine build_grid_arbitrary.  This routine
could not have been used because it had a hard-coded STOP call, and comments in
it indicated that it should have been deleted in July, 2013.  The run-time
parameter setting that would have triggered a call to this routine has been
retained for now, but with a fatal error message explaining that this routine
has not been implemented.   All answers are bitwise identical in any cases that
ran before.