Do not allocate ustar and tau_mag together
  Modified MOM_surface_forcing_gfdl.F90 and  MOM_surface_forcing.F90 to allocate
either ustar or tau_mag, but not both, in the forcing and mech_forcing types,
depending on whether the model is in Boussinesq mode.  Also added tests to
convert_IOB_to_forces in the FMS_cap code and to routines in MOM_surface_forcing
in the solo_driver code to ensure that only arrays that are associated are set.
All answers are bitwise identical, but checksum statements for the unused arrays
are eliminated when DEBUG = True.