Update default FMS to 2023.03
The default FMS build in ac/deps is updated to 2023.03.

FMS source now includes a suite of test programs which require explicit
preprocessing macros, which can complicate out makedep-based build when
those macros are not present.  To avoid this, the new "skip" flag has
been added to the makedep build.

The skip flag should not cause errors or other issues in older versions
of FMS which do not have the excluded directory (though perhaps that
could or should change in the future).